refactor: move AppointmentsScreen logic to its own file
