+++ src/presentation/screens/appointments-screen/appointments-screen.tsx
@@ -1,12 +1,11 @@
-import React, { useState, useEffect, useCallback } from 'react';
-import { useNavigation } from '@react-navigation/native';
-import { routes } from '~/presentation/navigation/routes';
+import React from 'react';
 
 // Components
 import { SafeAreaView } from 'react-native-safe-area-context';
 import { ActivityIndicator, FlatList } from 'react-native';
 import { AppointmentButtonCard } from '~/presentation/components/buttons';
 import { Heading } from '~/presentation/components/heading';
+import { ErrorScreen, EmptyList } from '~/presentation/components';
 
 // Styles
 import {
@@ -15,47 +14,21 @@ import {
   HeaderContainer,
   Footer,
 } from './styles';
-import { useTheme } from 'styled-components';
 
 // Types
 import { AppointmentModel } from '~/domain/models';
-import { makeRemoteGetAppointments } from '~/main/factories/usecases';
-import { ErrorScreen, EmptyList } from '~/presentation/components';
-import { StackNavigation } from '~/presentation/navigation/navigators/types';
+import { useAppointmentsScreenViewController } from './appointments-screen-view-controller';
 
 export function AppointmentsScreen() {
-  const [appointments, setAppointments] = useState<AppointmentModel[]>([]);
-  const [isLoading, setIsLoading] = useState(false);
-  const [hasError, setHasError] = useState(false);
-  const theme = useTheme();
-  const navigation = useNavigation<StackNavigation>();
-
-  const fetchAppointments = useCallback(async () => {
-    setIsLoading(true);
-    setHasError(false);
-    const remoteGetAppointments = makeRemoteGetAppointments();
-
-    try {
-      const response = await remoteGetAppointments.execute();
-      setIsLoading(false);
-      setAppointments(response.data);
-    } catch (error) {
-      setIsLoading(false);
-      setHasError(true);
-    }
-  }, []);
-
-  useEffect(() => {
-    void fetchAppointments();
-  }, [fetchAppointments]);
-
-  function goToAppointmentDetailsScreen(id: number) {
-    return navigation.navigate(routes.AppointmentDetailsScreen, { id });
-  }
-
-  function keyExtractor(item: AppointmentModel) {
-    return item.id.toString();
-  }
+  const {
+    appointments,
+    theme,
+    keyExtractor,
+    goToAppointmentDetailsScreen,
+    isLoading,
+    hasError,
+    fetchAppointments,
+  } = useAppointmentsScreenViewController();
 
   function renderItem({ item }: { item: AppointmentModel }) {
     return (
